devverse_ai_web_project/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
sqlalchemy
